refactor: move graphql query to own file
