+++ src/aioamazondevices/api.py
@@ -63,6 +63,7 @@ from .exceptions import (
     CannotRetrieveData,
     WrongMethod,
 )
+from .query import QUERY_DEVICE_STATE
 from .utils import obfuscate_email, scrub_fields
 
 
@@ -582,95 +583,7 @@ class AmazonEchoApi:
         self,
     ) -> dict[str, Any]:
         """Get Device State."""
-        query = """
-        query getDevicesState ($latencyTolerance: LatencyToleranceValue) {
-          listEndpoints(listEndpointsInput: {}) {
-            endpoints {
-              endpointId: id
-              friendlyNameObject { value { text } }
-              manufacturer { value { text } }
-              model { value { text} }
-              serialNumber { value { text } }
-              softwareVersion { value { text } }
-              creationTime
-              enablement
-              settings {
-                doNotDisturb {
-                  id
-                  endpointId
-                  name
-                  toggleValue
-                  error {
-                    type
-                    message
-                  }
-                }
-              }
-              displayCategories {
-                all { value }
-                primary { value }
-              }
-              alexaEnabledMetadata {
-                iconId
-                isVisible
-                category
-                capabilities
-              }
-              legacyIdentifiers {
-                dmsIdentifier {
-                  deviceType { value { text } }
-                }
-                chrsIdentifier { entityId }
-              }
-              legacyAppliance { applianceId }
-              associatedUnits { id }
-              connections {
-                  type
-                  macAddress
-                  bleMeshDeviceUuid
-              }
-              features(latencyToleranceValue: $latencyTolerance) {
-                name
-                instance
-                properties {
-                  name
-                  type
-                  accuracy
-                  error { message }
-                  __typename
-                  ... on Illuminance {
-                    illuminanceValue { value }
-                    timeOfSample
-                    timeOfLastChange
-                  }
-                  ... on Reachability {
-                      reachabilityStatusValue
-                      timeOfSample
-                      timeOfLastChange
-                  }
-                  ... on DetectionState {
-                      detectionStateValue
-                      timeOfSample
-                      timeOfLastChange
-                  }
-                  ... on Volume {
-                      value { volValue: value }
-                  }
-                  ... on TemperatureSensor {
-                      name
-                      value {
-                        value
-                        scale
-                      }
-                      timeOfSample
-                      timeOfLastChange
-                  }
-                }
-              }
-            }
-          }
-        }
-        """
+        query = QUERY_DEVICE_STATE
 
         payload = {
             "operationName": "getDevicesState",
